remove server.hpp
